Create 8 criminals problem
The criminal investigation brigade are now finding criminal for a criminal case. There are 8 suspects and here are the preliminary conclusions:

There is at least one criminal in A and B;
There are at least two criminals among A, C, D and E;
B and E may be both criminals or they are both innocent;
There are at least three criminals among C, D, F and G;
There is only one criminals in G and H;
If E is innocent, H is also innocent.
Write a program to find out there are how many possible combinations of eight suspects' positions.

Hint: each combination can be expressed by a list like [1,0,0,0,1,1,1,0], where the  𝑖𝑡ℎ  element of the list is 0 if the  𝑖𝑡ℎ  suspect is not a criminal, and the  𝑖𝑡ℎ  element of this list is 1 if the  𝑖𝑡ℎ  suspect is a criminal.